+++ Common/ValidDomain.h
@@ -3,6 +3,6 @@
 
 #include <stdbool.h>
 
-bool isValidDomain(const char * const domain, const int lenDomain);
+bool isValidDomain(const char * const domain, const size_t lenDomain);
 
 #endif
